fix/automatically play when no following file

+++ add-on/package.sh
@@ -7,9 +7,14 @@ if [ -f "$FILE" ]; then
     rm $FILE
 fi
 
+# Check if the file exists
+if [ -d "output" ]; then
+    echo "output exists, removing..."
+    rm -r output
+fi
+
 # remove unnecessory files
 find . | grep -E "(/__pycache__$|\.pyc$|\.pyo$)" | xargs rm -rf
-rm -r output
 
 # Zip all files excluding meta.json into anki_quick_ai.ankiaddon
 zip -r $FILE * -x "meta.json" -x "package.sh"
